bug 347974: Fix handling of previous value in Wires which have flow control filtering.
The fix "primes the pump" by setting the first value passed to the wire as
the previous value even if that value is not actually passed to the consumer.

The fix also makes the proper distinction between lastValue, which is the
last value provided by the Producer, and previousValue, which is the previous
value supplied to the Consumer. When wire flow control is in effect, these 2
values are often different since some values are not actually supplied to the
consumer.

Finally the fix adds finally blocks to properly set lastValue and/or
previousValue due to early return from the method.