Bug 362232 - [coordinator] Add support for the new, mandatory orphaned coordination requirement.
New CoordinationReferent. It is tracked by the CoordinationWeakReference in order to know when an initiator has abandoned all strong
references to the coordination without properly ending it. It is also used as a simple proxy when making a coordination available to clients
other than the initiator.

New CoordinationWeakReference. Maintains a strong reference to the CoordinationImpl. When it is placed on the ReferenceQueue, the
CoordinationImpl is retrieved and failed with Coordination.ORPHANED.

A CoordinationReferent is returned to callers of Coordinator.create() and begin(). Only instances returned by these methods are tracked by
CoordinationWeakReference. Other API methods that receive coordinations as results (e.g., Coordinator.getCoordinations()) or parameters (e.g.,
Participant.failed()), receive different instances.

Note that one of the goals was to minimize the impact of this change on existing code. Consequently, all internal data structures and methods
continue to use CoordinationImpl. It is only references to the initiator's instance of CoordinationReferent that must not be strongly
referenced by the implementation. When providing coordination instances externally to clients other than the initiator, CoordinationImpl is
wrapped with a new CoordinationReferent.

Coordinations will remain visible to callers of relevant methods (other than create() or begin()) until the orphaned coordination has been
removed by the fail process, even when CoordinationWeakReference.get() would return null. This makes the flow of orphaned coordinations
seamless with others. A management agent, for example, might detect that a coordination is obsolete and fail it before the orphan cleanup
process executes. CoordinationWeakReference instances are transparently tracked by the CoordinationWeakReference class.