Bug 362232 - [coordinator] Add support for the new, mandatory orphaned coordination requirement.
Modified CoordinationReferent.equals to return false if the object is not an instance of CoordinationReferent. In addition to the one listed
below, this change required a couple of other changes to CoordinationImpl methods that were still returning CoordinationImpl instead of
CoordinationReferent.

Updated the while loop for unwinding the thread local stack in CoordinationImpl.end to compare apples to apples.