Bug 362232 - [coordinator] Add support for the new, mandatory orphaned coordination requirement.
Removed the set of weak references maintained by CoordinationWeakReference. Each CoordinationImpl now maintains a strong reference to its
associated CoordinationWeakReference.

CoordinationImpl now maintains a CoordinationReferent that is a different instance than the one returned to the initiator. It is used to pass
the referent to clients other than the initiator without needing to create new objects.