Bug 362137: Support for coordination maximum timeout.
I added support for a new property named
org.eclipse.equinox.coordinator.timeout. The value is in milliseconds. The
default value is 0.

The property is retrieved using the coordinator's BundleContext.getProperty.
The property is retrieved anew with each instantiation of a Coordinator
service.

If a client attempts to create a coordination with a timeout less than the
maximum timeout (when max timeout != 0), the requested timeout will be
overwritten with the max timeout. If max timeout != 0, clients will not be able
to extend the timeout of a coordination beyond the max.