Bug 363374: Additional unresolved bundle tests.
There were a couple of places where BundleWiring.findEntries was still being used instead of Bundle.findEntries. These new tests are designed
to catch those in the future. In order to do that, the test had to be setup so that a resource bundle and icon was used, as well as ensure a
call to MetaTypeInformation.getLocales occurred.