Bug 366188 - Support for authorized_keys file
If the system property 'ssh.server.authorized_keys' is specified it

may point to a file which contains a list of keys that are allowed to

connect to the SSH console. The file will be read every time a

connection is made so that it can be modified dynamically at runtime.