Bug 348993: Allow to stop a thread from console (e.g. main thread that is caught in an endless loop)
The existing threads command was updated to support additional syntax. The full syntax is now

(t | threads) [stop <thread> [<throwable>]]

where <thread> is the thread name, and <throwable> is the fully qualified class name of something extending
java.lang.Throwable and visible to the console bundle's class loader. If <throwable> is unspecified, the default is
java.lang.IllegalStateException. The <throwable> will be created with a translatable message indicating the thread was
stopped by the OSGi Console, if a constructor accepting a String as the sole parameter exists. Otherwise, the default
constructor is used. The <throwable> is initialized with the target thread's stack trace and a cause of
java.lang.RuntimeException which will also contain the translatable message. The cause will contain the stack trace of
the current thread.