Bug 406936 - [rfc 191] Grant implied import package permissions to woven bundles with dynamic imports.
Check that the weaver has PackagePermission[pkg,IMPORT] when adding or
setting anything to the DynamicImportList.
Other modifications to the list currently do not require this permission
but still require AdminPermission[bundle,WEAVE].

Add dedicated method to BundlePermissions for adding permissions to
woven bundles. A PackagePermission with an action of
IMPORT must be used. These permissions are tracked in their own
Permissions collection and are now part of the implies check.

WovenClassImpl grants the permissions right after transitioning to
TRANSFORMED but before adding the imports to the
BundleLoader. The ProtectionDomain is gotten by casting the bundle
wiring revision to ModuleRevision and then the revision
info to Generation. The permissions of the protection domain are then
cast to BundlePermissions.