Bug 405919 - [unity] support for osgi.checkConfiguration option
Update EquinoxConifguration to support the osgi.checkConfiguration property.

Update BundleInfo.Generation to support a lastModified attribute. The value is the
last modified timestamp of either the bundle manifest (if a directory) or the bundle jar
at installation time.

Update Storage to load and store the last modified data. Whether or not a bundle
needs discarding is also computed here.