Bug 405918 - [unity] save persistent data after period of inactivity
Add the eclipse.stateSaveDelayInterval property, along with its default value, to EquinoxConfiguration. The
configuration is responsible for filling in the default value and type validation. Zero means the state is
saved immediately with each update. A negative value means the state is never saved as part of an update. A
positive value results in a scheduled task being created that executes every value milliseconds and saves the
current state.

Add a ScheduledExecutorService to EquinoxContainer with a min pool size of zero and a max size of one. The service
is created on init() and shutdown on close().

Add the method updatedDatabase() to ModuleContainerAdaptor. This method is called whenever a database update occurs.
The default behavior is to do nothing. The EquinoxContainerAdaptor overrides the method and saves the database state
based on the value of the eclipse.stateSaveDelayInterval property.

Add the StorageSaver class. It encapsulates the scheduling of the task, registering of the shutdown hook, and determining
which type of save is required. This same functionality was removed from EquinoxContainerAdaptor. EquinoxContainer
initializes the StorageSaver on init() and disposes of it on close(). EquinoxContainerAdaptor receives the updatedDatabase()
call from ModuleDatabase and requests a save by getting the StorageSaver from the container.

The osgi.framework.activeThreadType property is obsolete.

The framework active thread task is no longer necessary because the ExecutorService maintained by the
EquinoxContainer will always have one and only one active, non-daemon thread until system shutdown.