[Bug 415461] The constructor ApplicationAdminPermission.SignerWrapper(String) is never used locally.
This warning comes from OSGi source code included in the bundle. We do not wish to modify this source so
a warning suppression was added to the pom.xml, just as it had previously been added to build.properties
under the old build structure (see bug 266000).

Update bundle version.

Update copyright.