[Bug 415461] The serializable class does not declare a static final serialVersionUID field of type long.
Did not wish to change line 223 as the current way of constructing the ArrayList seemed crucial to the test.
Could not use @SuppressAnnotations("serial") because the compiler compliance level is set to 1.4. We can change this
but it would lead to the need for fixing a number of new warnings.

Update copyright.