Bug 416640 - [metatype] Consider providing API access to min/max values from metadata XML.
Add new methods to the EquinoxAttributeDefinition interface allowing access to the min and max values. The return type is
String which matches the type defined by the metatype XML schema.

Update implementation to accept min/max values that are not integers for type String. If min and max are integers, a string
length comparison takes place. If either min or max is not an integer, then String.compareTo is called.

Increment bundle minor version.

Update equinox metatype api version to 1.3.