Bug 419141 - Issues with reading the wiring graph without holding locks
Differentiate invalidating wiring vs. unloading it.  When unloading we need to make sure the ModuleLoader is freed for GC.  When invalidating we don't need to null out the ModuleLoader because the ModuleWiring is getting tossed anyway.

Do null checks against the ModuleLoader to cover the case where a wiring becomes invalid.