Bug 421487 - Check for null (again) when unwinding the thread local coordination stack.
The previous fix still had a small window of opportunity for an NPE to occur. Each call to peek results in orphaned
coordination processing, so despite the initial check for coordinator.peek() != null, the subsequent calls to peek() could
result in null if a coordination on the thread local stack became orphaned during that time.