[Bug 420067] Add support for org.eclipse.osgi/resolver/providers resolver trace option.
Specifying org.eclipse.osgi/resolver/providers = true within the .options file will enable trace output describing what's
happening between the resolve context and resolver when finding providers. A list of the current output follows.

(1) The requirement for which providers are being found.
(2) Capabilities filtered because their effective directive was not equal to "resolve".
(3) Capabilities filtered by resolver hooks via the filterMatches method.
(4) Capabilities filtered because their resource was no resolved.
(5) Capabilities filtered because the requirer did not have permission.
(6) Capabilities filtered because the provider did not have permission.
(7) Capabilities filtered because they were disabled.
(8) Capabilities filtered because they were substituted.
(9) The list of capabilities being returned to the resolver after filtering.

I tried to format the output so that it's easily readable using line separators and tabs. The risk, of course, is that while
it might look good in the console output, it could look horrible when System.out is routed to some third-party logging
mechanism.