Bug 436758 - Installing a bundle that requires Java 7 when running 6 with -clean cannot recover
 - Fixed by seeding the module database timestamps with the current
time.  This also impacted the check when loading persistent data into
the database.  There is a check to ensure loading persistent data can
only happen on an empty database.  The check used to check for zero
timestamp value.  This is never going to be true.  I updated this to
check for construction time instead.

Change-Id: Ic5a79577476c6890235219aae829574f0060e385
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>