Bug 444201 - Make embedded JVM work again on MacOS
With the new native MacOS file system layout (bug 431116) any embedded JVM
will fail to start.  Reason is a too simple detection of the JVM bundle
based on the existence of a /Contents/ path segment, which now happens to
always be there and causes the JVM to be loaded at the Eclipse.app/ location
instead of at the embedded JVM bundle location.  Fix is to check for
/Contents/Home/ which should be unique enough for MacOS JVM bundles.

This also fixes bug 442324's issue of a wrongly required legacy Apple JVM even
if a JVM is embedded.

Bug: 442324
Bug: 444201
Change-Id: I72b55b1f6dfdbd9d08db605308a2ec4456edf858
Signed-off-by: Christian Georgi <christian.georgi@sap.com>