Bug 441421 - Eclipse crashes with UnsatisfiedLinkError on ppc64le
The code change is to set SWT_GTK3 environment variable to 0 in case of
linux on ppc64le architecture.

Since this change is required only on linux and ppc64le platform, the
code changes are under LINUX and __ppc64le__ preprocessor macros. 

Added a new preprocessor macro __$(architecture)__ to make_linux.mak so
that it can be used in the compilation of eclipse.c

Please note: __ppc64le__ is not available as predefined macro in gcc
that we are using for build
Change-Id: I9748254e6ebdb9bd49f88726657171f338c6dc1a
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>