Bug 470527 - Replace `new Boolean` with `Boolean.valueOf`
Using `new Boolean()` results in the creation of a new object on the
heap, when the flyweight `Boolean.TRUE` and `Boolean.FALSE` are available.
Java 1.4 added a `Boolean.valueOf()` which can be used in place of
`new Boolean()` but which will use the existing flyweight values instead.

Globally change `new Boolean(...)` to `Boolean.valueOf(...)` and replace
constant valued expressions with their flyweight counterparts.

Defer changes to `org.osgi.util.xml.XMLParserActivator` to upstream OSGi.

Bug: 470527
Change-Id: I122e4456a49e8eee9d88db0a3237f7e6b19a8469
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>