Bug 475416 - Add timing information to BundleLoader
When debugging slow start times, much of the slowness can be
attributed to classes that are slow to be defined, typically
due to dependencies on other bundles' activators being executed.

By calculating the time taken, hot spots can be identified
and resolved.

Timing information is only displayed if the debug option
debug/loader is enabled.

Bug: 475416
Change-Id: I5605fa2c2eb82d0b79d1af086d9e0a8bd69d5867
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>