Bug 475924 - Measure activator class loading at bundle start
In order for a bundle to be started, the appropriate BundleActivator
must be located, then executed with a call to start(). However, the
lookup of the class itself may take some time, with the result that
the overall loading time of the bundle is not accurately represented.

By moving the start of the debug loading time to the start of the
method, the lookup for the BundleActivator class can be measured as
well as the call to start().

Change-Id: I6e69cdc7109bfb90d3b9bb61c52171a4fad6f900
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>