Bug 487954 - SignedBundleHook.getSignedContent() should report path
My guess is the "root" problem is in the JRE, but since they are deep in
native methods, they may know the "file name" at the point the error
occurs. But, we do, so I propose to catch the ZipException, and
supplement it with the path information, and then re-throw it. I think
that minimizes risk of effecting other code paths. 


Change-Id: Ieaf0072bb06f8be541a52cc24e35c38a82985e27
Signed-off-by: David Williams <david_williams@us.ibm.com>
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>