Bug 487842 - Unable to acquire state change lock for the module when starting a bundle
Must increment inStart BEFORE acquiring the STARTED state the first
time.  There is still an unavoidable timing issue.  The window is now
as small as possible.  A new flag is added to disable auto-starting
newly resolved bundles that are marked for persistent start.

Also fixed bug in how modules are restarted after a refresh.
They must use their activation policy if it is set.

Change-Id: I8175af9ed29e611659a8b8bc0b64b0a3874e68bf
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>