[Bug 477787] Embedded Equinox in an entreprise application ear - NullPointerException at start up
In BundleInfo.Generation, check for null content and, if so, set lastModified to zero. In ModuleClassLoader, check for a null base file from BundleFile and, if so, create the
ProtectionDomain with a null CodeSource. In both cases, the referenced variable will be null when the osgi.framework property was specified with an invalid value.

Plus test.

Signed-off-by: John Ross <jwross@us.ibm.com>