Bug 490008 - Avoid storing complete jar Manifest object
Storing a whole Manifest for what is essentially minimal data is very
space intensive. Not only does this result in an additional 1/2Mb of
string data, the other data structures (particularly HashMaps) store dby
Manifest objects ends up being able to reduce the start-up of a generic
Eclipse SDK instance from a minimum of around 44Mb to 39Mb after
performing repeated GC operations.

The custom data structures are not necessarily as optimal as they could
be (for example, repeated -Vendor strings aren't interned or uniquifed)
but if nothing is present then a singleton NONE instance is used and no
HashMap is instantiated. For an initial implementation it does well.

Change-Id: I7e0a049a3cfeb42f49695576bb2feb71be8700ed
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>