Bug 496034 - Obtain readLock on ModuleDatabase before classLoaderMonitor
To prevent potential out of order locks between the ModuleDatabase 
read/write lock and the classLoaderMonitor lock we obtain the readLock
on the ModuleDatabase before the classLoaderMonitor.  This is because
during the refresh operation the removed wirings will close the
BundleLoader while holding the writeLock, this eventually nests a lock
with the classLoaderMonitor monitor lock.  Since we call foreign hook
code while holding the classLoaderMonitor monitor lock we should be
consistent in the order locks are obtained to prevent potential
deadlocks by the hooks.

Change-Id: I0987756b2046ca8d8076bc7ad01a2cc9e87e18ff
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>