Bug 497092 - Use INFO event from uninstalled/updated bundle revision
If trying to open a zip file and an IOException is thrown then check if
the bundle revision still exists in the list of revisions for the
bundle.  If it does not then the revision was removed as part of an
uninstall or update operation.  In this case just use an INFO event.

Change-Id: I975061a2e963be1c7b9752401aae6d70eddda003
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>