Bug 510673 - Set ds.delayed.keepInstances true by default
This is done in code and as a p2 instruction.  The p2 instruction is a
safe guard when provisioned with p2 so that we ensure the property is
set in the config.ini such that it will be enabled at boot with no
chance of being missed by felix SCR if it happens to start before
equinox.ds.

Change-Id: I66dbec7b33b14bbc85e8d8d64d71e930106c415e
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>