Bug 510746 - Test failures in official build in equinox.http.servlet.tests: ServletTest#test_WBServlet2
Equinox DS had a bug that allowed Character types to be represented as
single character strings.  The specification requires the Character to
be represented as an integer (seems bad for readability).  The felix SCR
implementation is compliant with the specification here.

I converted the test to use String instead so I could keep the XML more
readable instead of putting the int representation for 'b' in.

Change-Id: I1982397b2c9bb2a7a42a1e6cfd8f709c629c52f8
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>