Bug 517013 - Avoid memcpy@GLIBC_2.14 dependency on Linux x86_64
On newer Linux x86_64 hosts like REL7, an optimized version of 
memcpy() is available in glibc and linked in by default. The
drawback of linking that optimized version is, that the resulting
binaries don't run any more on older Linux distributions like
REL6 (where that optimized memcpy@GLIBC_2.14 is not available).

This fix avoids the problem by replacing memcpy() with memmove()
on Linux x86_64 hosts only. To be safe, it also compiles with
-fno-builtin-memcpy -fno-builtin-memmove on Linux x86_64. See
also bug 515155 for a very similar fix in SWT to support REL6.

Change-Id: I39bc8d4975015eb52f35d3d02eae921bf6d66478
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>
Signed-off-by: Arun Thondapu <arunkumar.thondapu@in.ibm.com>