Bug 521542 - org.osgi.framework.storage is used even when osgi.configuration.area is set
Log a warning when this occurs.

Had to add an exceptions map that can be used during bootstrap
when there is no log service to log to yet.  This allowed for 
fixing several TODOs where we thought we should be logging but
couldn't because no log service was available yet.