Bug 532194 - Use BundlePermissions for the EvaluationCacheKey
Also check for null Conditions from a row and update testcase to handle
when multiple conditions are used in a single row. If there is a
mix of mutable and immutable conditions then the satisfied conditions
are nulled out to avoid calling them again on future permission
checkes.  The method checkMutable needs to handle the null values.

Change-Id: I9adf9cf54bfe571d53049d48ffb44e76bc78f121
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>