Bug 386377 - Native launcher not built during first build
Change where the build.xml is invoked for building the native parts
during a maven build.

Previously, because the build.xml was invoked from the feature, and
the reactor orders the build of the native fragments *before* the
feature, it means the fragments do not include the freshly built
native parts.

Invoke it instead from the fragments themselves by moving the
profile to the launcher-binary-parent pom. This way the native
parts are guaranteed to be built in time to be included in the
fragments.

Change-Id: I0f9004cb10427cdd6a84bdf94cbbf10263f5267f
Signed-off-by: Mat Booth <mat.booth@redhat.com>