Bug 535463 - Fixes for running tests on Java 9
testContextFinderEmptyGetResources - converted this to a unit test for
BundleLoader.compoundEnumerations() method to take null parameters.
Reproducing the situation from the original testcase is not obvious,
decided to instead directly unit test the compondEnumerations to ensure
it can take null parameters without NPEs.

testContextFinderGetResource - moved this to SystemBundleTests which
allows us to set the parent class loader of the ContextFinder in order
to allow it to search resources in the framework instead of from the VM.

Change-Id: I9187dbda840afc1e8d7b17034854c770b7d16708
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>