Bug 548075 - Ability to control if bundles are activated in parallel
Add a new option equinox.start.level.restrict.parallel
which is false by default.  Setting this value to true
requires that the method Module.setParallelActivation must
be called with true to enable parallel activation for the
Module.

The parallel activation setting is persistent across restarts. Tests
have been added to test for persistence of the setting and that the
setting is applied correctly when activating bundles from the
start-level implementation.

Change-Id: I20ea5807edfd10762fc5758317d946946299d5a7
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>