Bug 549406 - improve performance of checkSystemBundle
Storage.checkSystemBundle is expensive because it has to reparse the
system bundle manifest and constructs a new ModuleRevision based on the
current environment and then does a deep equality check of the newly
constructed ModuleRevision with the cached ModuleRevision for the system
bundle.

This commit changes the implementation to instead base the check off the
values for the environments packages and capabilities compared to the
cached values of these from last start.  It also does a timestamp check
to see if the framework's content has actually changed.

Change-Id: I47a310dbf8f45acf46a95d9ffbfc79801118f5c0
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>