Bug 558371 - Avoid use of MethodHandle
Use of method handles does not seem to improve performance.  In my
testing it actually makes things slower on Java 8.  Particularly the use
of MethodHandle for calling the constructor for the value to match
against.

Substrate also doesn't like the use of MethodHandles in the
FilterImpl.  For now I think we should avoid using them unless they show
a clear performance win.

Change-Id: I220cb2cde0a7c5bf948db6e34f8d337e3a4ce0f8
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>