filter: Improvements to FilterImpl
The implementation is updated to better exploit the String type of
value. We don't need to pass it as a parameter since it is a final
field in the object. We also don't need to cast it to String since
it is already known to be a String. This then leads to refactoring
of the FilterImpl sub-type hierarchy.

Change-Id: I57c94dd772f776fb6d7cbbbc12eb3d3f129c6386
Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>