Bug 559118 - Don't fail with runtime exception when substitutes are not handled correctly
This doesn't impact the framework, but the resolver is also registered
as a service and used by others. The original fix would cause a runtime
exception if no sources could be found for a blamed used capability.
The previous resolver implementation would silently ignore the source
and it would not impact the used class space.  This still produces a
valid class space according to what the resolve context is giving the
resolver to work with.

We will just log an info message if this happens.

Change-Id: I34c61efa33a703c66763197146d09a62f09f9ad8