Bug 560428 - Mutex in the Windows launcher potentially gets not released
If a starting instance (with files to be opened on the command line)
calls takeDownSplash() (through JNI) before the timer in the native
launcher has fired for the first time, the associated Mutex never gets
released and therefore this instance will never be found and reused.
Pure E4 RCPs are likely to hit this one second time slot.

Change-Id: I744b5d3741bf7bc6ece360527f3962dd2d9d5081
Signed-off-by: Gerhard Kreuzer <gerhard.kreuzer@siemens.com>