Bug 563304 - close event manager on STOPPED_SYSTEM_REFRESHED
The STOPPED_SYSTEM_REFRESHED event was new in the OSGi R7 specification
but the code listening for framework STOPPED events to close the event
manager was not updated to pay attention to the new event.  The result
is that when this new event is used during framework shutdown the event
manager is not closed, causing a thread to remain active after framework
stop.

Change-Id: Ib1a1c360ede7ad1b733556bc2ce44c5469c984b2
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>