Bug 563372 - Be sure to handle async update properly
testBug258209_1 has a timing issue because the Framework.update
operation is async.  This allows the test to try to waitForStop but the
framework may have already be re-activated. Update this test to do what
other tests that update the framework do by calling waitForStop in
another thread before invoking Framework.update.

Change-Id: Iee4de68f05b7cb151be2f61fcad353a625eb455c
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>