Bug 564258 - [ds.tests] Fail in maven because of internal classes
Use BundleContext.getAllServiceReferences and ServiceTracker.open(true)
to indicate that the test wants to get all service references regardless
of any class space wiring issues.  This is to work around a build issue
with the tests that ends up packaging the types from the test bundles
into the main ds.tests bundle.

Change-Id: I7fd67f209f76e90e18001a1f24e064aed49586a1
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>