Bug 565276 - Add documentation for sneakyThrow to ServiceCaller
When using `ServiceCaller` to invoke a consumer that throws a
checked exception, the exception itself is not trivially thrown back
again. However, using the sneakyThrow technique, it is possible to fool
the compiler into throwing a checked exception as an unchecked
exception.

Document this case in the JavaDoc so that others who experience a
similar pattern can implement it in a similar way. This isn't API
on `ServiceCaller` because generally the implementation is not related
to this particular class and should not be used as such.

Change-Id: I48553ccea9d14f6ea852e8ca4d9a6fb87b129e92
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>