Bug 565522 - EquinoxClassLoader returns wrong resource content
When searching for a resource that is indexed by a bundle URL the search
must not return an entry unless the index properly matches the class
path index.

This issue is surfaced when taking the external String from a bundle
resource URL and creating a new URL from it.  Here the index is stored
in the port of the URL and it must be validated when searching for the
entry when opening the connection to the new URL

Change-Id: I74685811c39ba10af2c92851759cd1eb6ffcf339
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>